finalizado hasta análisis no supervisado

+++ advanced_machine_learning/examenes/examen_certificacion_intento_1_human_activity_recognition/notebooks/utils/kaggle_utils.py
@@ -6,9 +6,12 @@ import os
 import opendatasets as od
 
 
-def descargar_dataset_kaggle(url, camino_credenciales, data_dir='.'):
-    if os.path.exists("./kaggle.json"):
-        os.remove("./kaggle.json")
-    os.symlink(camino_credenciales, "./kaggle.json")
+def descargar_dataset_kaggle(url, camino_credenciales=None, data_dir='.'):
+    if camino_credenciales is not None:
+        if os.path.exists("./kaggle.json"):
+            os.remove("./kaggle.json")
+        os.symlink(camino_credenciales, "./kaggle.json")
     od.download(url, data_dir=data_dir)
-    os.remove("./kaggle.json")
+
+    if camino_credenciales is not None:
+        os.remove("./kaggle.json")
